requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==5.2.1
WebTest==2.0.33
factory-boy==2.12.0
pdbpp==0.10.2

# Lint and code style
black==19.3b0
flake8==3.7.8
flake8-blind-except==0.1.1
flake8-debugger==3.2.0
flake8-docstrings==1.5.0
flake8-isort==2.7.0
isort==4.3.21
pep8-naming==0.8.2


#virtualcam
Pillow==6.1.0

# for bootstrap macro
Bootstrap-Flask==1.1.0